LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/column-updates-computed-style-expected.txt


FAIL ::column updates computed style on its originating element assert_equals: expected "rgb(255, 0, 0)" but got ""
